runner/requirements.txt
dnspython==1.16.0
mysql-connector-python==8.0.19
protobuf==3.6.1
python-dotenv==0.12.0
schedule==0.6.0
six==1.14.0
speedtest-cli==2.1.2
